Parse MeteorDensity attribute as double instead of int
Fixes crash on loading some maps

+++ Mappy/Models/MapAttributesResult.cs
@@ -36,7 +36,7 @@ namespace Mappy.Models
 
         public int MeteorDuration { get; set; }
 
-        public int MeteorDensity { get; set; }
+        public double MeteorDensity { get; set; }
 
         public int MeteorInterval { get; set; }
